+++ bindings/go/scip/sort_test.go
@@ -77,7 +77,7 @@ func TestSortOccurrences(t *testing.T) {
 }
 
 func TestSortRanges(t *testing.T) {
-	occurrences := []*Range{
+	occurrences := []Range{
 		NewRange([]int32{2, 3, 5}),       // rank 2
 		NewRange([]int32{11, 10, 12}),    // rank 10
 		NewRange([]int32{6, 3, 5}),       // rank 4
@@ -93,7 +93,7 @@ func TestSortRanges(t *testing.T) {
 		NewRange([]int32{12, 4, 13, 8}),  // rank 12
 		NewRange([]int32{1, 3, 3, 5}),    // rank 1
 	}
-	unsorted := make([]*Range, len(occurrences))
+	unsorted := make([]Range, len(occurrences))
 	copy(unsorted, occurrences)
 
 	ranges := [][]int32{}
@@ -120,27 +120,5 @@ func TestSortRanges(t *testing.T) {
 	}
 	if diff := cmp.Diff(expected, ranges); diff != "" {
 		t.Errorf("unexpected occurrence order (-want +got):\n%s", diff)
-	}
-}
-
-func TestComparePositionToRange(t *testing.T) {
-	testCases := []struct {
-		line      int32
-		character int32
-		expected  int
-	}{
-		{5, 11, 0},
-		{5, 12, 0},
-		{5, 13, -1},
-		{4, 12, +1},
-		{5, 10, +1},
-		{5, 14, -1},
-		{6, 12, -1},
-	}
-
-	for _, testCase := range testCases {
-		if cmpRes := comparePositionToRange(5, 11, 5, 13, testCase.line, testCase.character); cmpRes != testCase.expected {
-			t.Errorf("unexpected ComparePositionSCIP result for %d:%d. want=%d have=%d", testCase.line, testCase.character, testCase.expected, cmpRes)
-		}
 	}
 }